REV (5) 2차 코드리뷰내용 반영 * 오타 제거 * 프로퍼티 은닉화를 위해 프로토콜 레벨에서의 저장프로퍼티 제거 * 프로토콜에 맞는 연산 프로퍼티 위치 재설정 (Shapable 프로퍼티에서 제거) * ErrorCode enum을 별도 파일로 분리 (공통 에러코드이므로)
