+++ CoordinateCalculator/CoordinateCalculator/Linable.swift
@@ -9,6 +9,6 @@
 import Foundation
 
 protocol Linable: Shapable {
-    var pointA : MyPoint {get}
-    var pointB : MyPoint {get}
+    var myLineInformation : Double {get}
 }
+
